wiiuseJ 0.12 work in progress. Added method to force stack type and fixed IR.z value and nunchuk thresholds.

+++ WiiUseJ/src/wiiusej/wiiusejevents/utils/EventsGatherer.java
@@ -103,7 +103,7 @@ public class EventsGatherer {
 	 * @param distance
 	 *            Pixel Distance between first two dots.
 	 */
-	public void prepareIRevent(int x, int y, int z, int ax, int ay, int xVRes,
+	public void prepareIRevent(int x, int y, float z, int ax, int ay, int xVRes,
 			int yVRes, int xOffset, int yOffset, short sensorBarPostion,
 			short screenAsPectRatio, short irSensitivity, float distance) {
 		genericEvent.prepareIRevent(x, y, z, ax, ay, xVRes, yVRes, xOffset,
